Draw subgraph.
