+++ src/main/java/nl/tudelft/context/model/graph/Node.java
@@ -1,5 +1,6 @@
 package nl.tudelft.context.model.graph;
 
+import nl.tudelft.context.controller.DefaultGraphController;
 import nl.tudelft.context.controller.GraphController;
 import nl.tudelft.context.controller.MainController;
 import nl.tudelft.context.drawable.DefaultLabel;
@@ -133,7 +134,7 @@ public class Node extends DefaultNode {
     }
 
     @Override
-    public DefaultLabel getLabel(final MainController mainController, final GraphController graphController,
+    public DefaultLabel getLabel(final MainController mainController, final DefaultGraphController graphController,
                                  final DrawableNode drawableNode) {
         return new InfoLabel(mainController, graphController, drawableNode, this);
     }
